Cross validation seems to be working

+++ mis_dro/epsilon.py
@@ -4,7 +4,7 @@ def get_num_observations_in_train_split(n_splits: int, split_idx: int, n_observa
     ratio = float(n_observations) / float(n_splits)
     num_splits_with_less_than_max_test_size = n_splits * np.ceil(ratio) - n_observations
     if split_idx < n_splits - num_splits_with_less_than_max_test_size:
-        return n_observations - np.ceil(ratio)
+        return int(n_observations - np.ceil(ratio))
     else:
-        return n_observations - np.floor(ratio)
+        return int(n_observations - np.floor(ratio))
